Add findCharacterLocation and findSteps methods to StepFinder Module
